+++ main.go
@@ -1,26 +1,28 @@
 package main
 
-func WokEvaluate(source string) WokData {
-	tokenizer := MakeTokenizer()
+import "woklang/woklang"
+
+func WokEvaluate(source string) woklang.WokData {
+	tokenizer := woklang.MakeTokenizer()
 	tokenizer.LoadFromString(source)
 	tokenizer.Tokenize()
 
-	parser := MakeParser()
+	parser := woklang.MakeParser()
 	parser.Parse(tokenizer.tokens)
 
-	interpreter := MakeInterpreter()
+	interpreter := woklang.MakeInterpreter()
 	return interpreter.Interpret(parser.expressions)
 }
 
 func main() {
 
-	tokenizer := MakeTokenizer()
+	tokenizer := woklang.MakeTokenizer()
 	tokenizer.LoadFromFile("demo.lisp")
 	tokenizer.Tokenize()
 
-	parser := MakeParser()
+	parser := woklang.MakeParser()
 	parser.Parse(tokenizer.tokens)
 
-	interpreter := MakeInterpreter()
+	interpreter := woklang.MakeInterpreter()
 	interpreter.Interpret(parser.expressions)
 }
