+++ woklang/runtime.go
@@ -1,4 +1,4 @@
-package main
+package woklang
 
 import "fmt"
