+++ woklang/scope.go
@@ -1,4 +1,4 @@
-package wokdata
+package woklang
 
 // Scope holds the structure of the environment of the runtime data state
 type Scope struct {
