+++ woklang/tokenizer.go
@@ -1,4 +1,4 @@
-package tokenizer
+package woklang
 
 import (
 	"fmt"
@@ -22,7 +22,7 @@ type Tokenizer struct {
 	source  []byte
 	current int
 	start   int
-	tokens  []Token
+	Tokens  []Token
 }
 
 func MakeTokenizer() Tokenizer {
@@ -35,12 +35,12 @@ func (tokenizer *Tokenizer) LoadFromFile(fileName string) {
 		tokenizer.Error(err.Error())
 	}
 	tokenizer.source = content
-	tokenizer.tokens = make([]Token, 0)
+	tokenizer.Tokens = make([]Token, 0)
 }
 
 func (tokenizer *Tokenizer) LoadFromString(source string) {
 	tokenizer.source = []byte(source)
-	tokenizer.tokens = make([]Token, 0)
+	tokenizer.Tokens = make([]Token, 0)
 }
 
 func (tokenizer *Tokenizer) Eof() bool {
@@ -79,7 +79,7 @@ func (tokenizer *Tokenizer) PeekNext() rune {
 }
 
 func (tokenizer *Tokenizer) AddToken(ttype TokenType, literal string) {
-	tokenizer.tokens = append(tokenizer.tokens, MakeToken(ttype, literal))
+	tokenizer.Tokens = append(tokenizer.Tokens, MakeToken(ttype, literal))
 }
 
 func (tokenizer *Tokenizer) Error(errorMessage string) {
